Put ImmutableCommonModel in common.py
