+++ ansible_base/lib/abstract_models/__init__.py
@@ -1,5 +1,5 @@
-from .common import CommonModel, NamedCommonModel, UniqueNamedCommonModel
-from .immutable import ImmutableCommonModel
+from .common import CommonModel, ImmutableCommonModel, NamedCommonModel, UniqueNamedCommonModel
+from .immutable import ImmutableModel
 from .organization import AbstractOrganization
 from .team import AbstractTeam
 
@@ -7,6 +7,7 @@ __all__ = (
     'AbstractOrganization',
     'AbstractTeam',
     'CommonModel',
+    'ImmutableModel',
     'ImmutableCommonModel',
     'NamedCommonModel',
     'UniqueNamedCommonModel',
